9219: cleanup
